Move `PlatformContextTextInputService` to a separate file.
