+++ compose/ui/ui/api/desktop/ui.api
@@ -3330,10 +3330,6 @@ public abstract interface class androidx/compose/ui/platform/PlatformContextText
 	public fun updateTextLayoutResult (Landroidx/compose/ui/platform/TextFieldStateAdapter;Landroidx/compose/ui/text/input/OffsetMapping;Landroidx/compose/ui/text/TextLayoutResult;Lkotlin/jvm/functions/Function1;Landroidx/compose/ui/geometry/Rect;Landroidx/compose/ui/geometry/Rect;)V
 }
 
-public final class androidx/compose/ui/platform/PlatformContext_skikoKt {
-	public static final fun asPlatformTextInputService (Landroidx/compose/ui/platform/PlatformContextTextInputService;)Landroidx/compose/ui/text/input/PlatformTextInputService;
-}
-
 public abstract interface class androidx/compose/ui/platform/PlatformDragAndDropManager {
 	public abstract fun drag-12SF9DM (Landroidx/compose/ui/draganddrop/DragAndDropTransferData;JLkotlin/jvm/functions/Function1;)Z
 	public abstract fun getModifier ()Landroidx/compose/ui/Modifier;
