+++ compose/ui/ui/src/skikoMain/kotlin/androidx/compose/ui/platform/PlatformContext.skiko.kt
@@ -299,140 +299,3 @@ internal class DelegateRootForTestListener : PlatformContext.RootForTestListener
     }
 }
 
-/**
- * Platform specific text input service.
- *
- * This is a non-deprecated version of [PlatformTextInputService], which is needed because we can't
- * expose deprecated APIs in [PlatformContext].
- */
-interface PlatformContextTextInputService {
-    /**
-     * Start text input session for given client.
-     *
-     * @see TextInputService.startInput
-     */
-    fun startInput(
-        value: TextFieldStateAdapter,
-        imeOptions: ImeOptions,
-        onEditCommand: (List<EditCommand>) -> Unit,
-        onImeActionPerformed: ((ImeAction) -> Unit)?
-    )
-
-    /**
-     * Restart input and show the keyboard. This should only be called when starting a new
-     * `PlatformTextInputModifierNode.textInputSession`.
-     *
-     * @see TextInputService.startInput
-     */
-    fun startInput() {}
-
-    /**
-     * Stop text input session.
-     *
-     * @see TextInputService.stopInput
-     */
-    fun stopInput()
-
-    /**
-     * Request showing onscreen keyboard
-     *
-     * There is no guarantee nor callback of the result of this API.
-     *
-     * @see TextInputService.showSoftwareKeyboard
-     */
-    fun showSoftwareKeyboard()
-
-    /**
-     * Hide software keyboard
-     *
-     * @see TextInputService.hideSoftwareKeyboard
-     */
-    fun hideSoftwareKeyboard()
-
-    /**
-     * Notify the new editor model to IME.
-     *
-     * @see TextInputSession.updateState
-     */
-    fun updateState(oldValue: TextFieldStateAdapter?, newValue: TextFieldStateAdapter)
-
-    /**
-     * Notify the focused rectangle to the system.
-     *
-     * The system can ignore this information or use it to for additional functionality.
-     *
-     * For example, desktop systems show a popup near the focused input area (for some languages).
-     */
-    // TODO(b/262648050) Try to find a better API.
-    fun notifyFocusedRect(rect: Rect) {
-    }
-
-    /**
-     * Notify the input service of layout and position changes.
-     *
-     * @see TextInputSession.updateTextLayoutResult
-     */
-    fun updateTextLayoutResult(
-        textFieldValue: TextFieldStateAdapter,
-        offsetMapping: OffsetMapping,
-        textLayoutResult: TextLayoutResult,
-        textFieldToRootTransform: (Matrix) -> Unit,
-        innerTextFieldBounds: Rect,
-        decorationBoxBounds: Rect
-    ) {
-    }
-}
-
-@Suppress("DEPRECATION")
-fun PlatformContextTextInputService.asPlatformTextInputService() = object : PlatformTextInputService {
-    override fun startInput(
-        value: TextFieldValue,
-        imeOptions: ImeOptions,
-        onEditCommand: (List<EditCommand>) -> Unit,
-        onImeActionPerformed: (ImeAction) -> Unit
-    ) {
-        this@asPlatformTextInputService.startInput(
-            value = value.asTextFieldStateAdapter(),
-            imeOptions = imeOptions,
-            onEditCommand = onEditCommand,
-            onImeActionPerformed = onImeActionPerformed
-        )
-    }
-
-    override fun startInput() = this@asPlatformTextInputService.startInput()
-
-    override fun stopInput() = this@asPlatformTextInputService.stopInput()
-
-    override fun showSoftwareKeyboard() = this@asPlatformTextInputService.showSoftwareKeyboard()
-
-    override fun hideSoftwareKeyboard() = this@asPlatformTextInputService.hideSoftwareKeyboard()
-
-    override fun updateState(oldValue: TextFieldValue?, newValue: TextFieldValue) {
-        this@asPlatformTextInputService.updateState(
-            oldValue = oldValue?.asTextFieldStateAdapter(),
-            newValue = newValue.asTextFieldStateAdapter()
-        )
-    }
-
-    override fun notifyFocusedRect(rect: Rect) {
-        this@asPlatformTextInputService.notifyFocusedRect(rect)
-    }
-
-    override fun updateTextLayoutResult(
-        textFieldValue: TextFieldValue,
-        offsetMapping: OffsetMapping,
-        textLayoutResult: TextLayoutResult,
-        textFieldToRootTransform: (Matrix) -> Unit,
-        innerTextFieldBounds: Rect,
-        decorationBoxBounds: Rect
-    ) {
-        this@asPlatformTextInputService.updateTextLayoutResult(
-            textFieldValue = textFieldValue.asTextFieldStateAdapter(),
-            offsetMapping = offsetMapping,
-            textLayoutResult = textLayoutResult,
-            textFieldToRootTransform = textFieldToRootTransform,
-            innerTextFieldBounds = innerTextFieldBounds,
-            decorationBoxBounds = decorationBoxBounds
-        )
-    }
-}
